requirements.txt
einops==0.8.1
nibabel==5.3.2
numpy==1.26.4
Pillow==11.0.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
wandb==0.19.11
opencv-python==4.11.0.86
matplotlib==3.10.3
